Фикс загрузки избранного, очистки активного офера
