+++ src/components/header/header.tsx
@@ -3,8 +3,11 @@ import { AppRoute, AuthorizationStatus } from '../../const';
 import classNames from 'classnames';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { logoutAction } from '../../store/api-actions';
-import { getAuthorizationStatus } from '../../store/user-process/selector';
-import { UserProfile } from '../user-profile/user-profile';
+import {
+	getAuthorizationStatus,
+	getUserName,
+} from '../../store/user-process/selector';
+import { FavoriteCount } from '../favorite-count/favorite-count';
 
 type HeaderProps = {
 	withNavigation?: boolean;
@@ -14,6 +17,7 @@ function Header({ withNavigation = true }: HeaderProps): JSX.Element {
 	const dispatch = useAppDispatch();
 	const { pathname } = useLocation();
 	const authorizationStatus = useAppSelector(getAuthorizationStatus);
+	const userName = useAppSelector(getUserName);
 
 	return (
 		<header className="header">
@@ -39,7 +43,18 @@ function Header({ withNavigation = true }: HeaderProps): JSX.Element {
 						<nav className="header__nav">
 							{authorizationStatus === AuthorizationStatus.Auth ? (
 								<ul className="header__nav-list">
-									<UserProfile />
+									<li className="header__nav-item user">
+										<Link
+											className="header__nav-link header__nav-link--profile"
+											to={AppRoute.Favorites}
+										>
+											<div className="header__avatar-wrapper user__avatar-wrapper"></div>
+											<span className="header__user-name user__name">
+												{userName}
+											</span>
+											<FavoriteCount />
+										</Link>
+									</li>
 									<li className="header__nav-item">
 										<Link
 											className="header__nav-link"
