+++ src/pages/main-page/hooks/use-current-offers.ts
@@ -3,11 +3,9 @@ import { useAppDispatch, useAppSelector } from '../../../hooks';
 import { getActiveCity, getOffers } from '../../../store/offers-data/selector';
 import { ServerOffer } from '../../../types/offer';
 import { fetchOffersAction } from '../../../store/api-actions';
-import { offersActions } from '../../../store/offers-data/offers-data';
 
 export function useCurrentOffers() {
 	const dispatch = useAppDispatch();
-	dispatch(offersActions.setActiveOffer(null));
 	const offers = useAppSelector(getOffers);
 	const activeCity = useAppSelector(getActiveCity);
 	const offersByCities: Record<string, ServerOffer[]> = {};
@@ -21,7 +19,6 @@ export function useCurrentOffers() {
 		offersByCities[city] = [offer];
 		continue;
 	}
-
 	const currentOffers = offersByCities[activeCity];
 
 	useEffect(() => {
