+++ src/store/favorites-data/favorites-data.ts
@@ -9,6 +9,7 @@ import {
 
 const initialState: FavoritesData = {
 	favorites: [],
+	isIdle: true,
 	isFavoritesLoading: false,
 };
 
@@ -20,6 +21,7 @@ export const favoritesData = createSlice({
 		builder
 			.addCase(fetchFavoritesAction.pending, (state) => {
 				state.isFavoritesLoading = true;
+				state.isIdle = false;
 			})
 			.addCase(fetchFavoritesAction.fulfilled, (state, action) => {
 				state.isFavoritesLoading = false;
